feat(front): check update at startup
